Reformatted KGGLM Trainer to inherit from Trainer designed for HuggingFace models and implemented PEARLM
